Fix import structure of ertwidgets
