+++ src/ert/gui/ertwidgets/models/__init__.py
@@ -1,5 +1,15 @@
+from .activerealizationsmodel import ActiveRealizationsModel
+from .path_model import PathModel
+from .selectable_list_model import SelectableListModel
+from .targetensemblemodel import TargetEnsembleModel
 from .text_model import TextModel
+from .valuemodel import ValueModel
 
 __all__ = [
     "TextModel",
+    "ActiveRealizationsModel",
+    "ValueModel",
+    "TargetEnsembleModel",
+    "SelectableListModel",
+    "PathModel",
 ]
