+++ src/ert/gui/simulation/evaluate_ensemble_panel.py
@@ -5,10 +5,12 @@ from qtpy import QtCore
 from qtpy.QtWidgets import QFormLayout, QLabel
 
 from ert.gui.ertnotifier import ErtNotifier
-from ert.gui.ertwidgets.copyablelabel import CopyableLabel
-from ert.gui.ertwidgets.ensembleselector import EnsembleSelector
-from ert.gui.ertwidgets.models.activerealizationsmodel import ActiveRealizationsModel
-from ert.gui.ertwidgets.stringbox import StringBox
+from ert.gui.ertwidgets import (
+    ActiveRealizationsModel,
+    CopyableLabel,
+    EnsembleSelector,
+    StringBox,
+)
 from ert.gui.simulation.experiment_config_panel import ExperimentConfigPanel
 from ert.mode_definitions import EVALUATE_ENSEMBLE_MODE
 from ert.run_models.evaluate_ensemble import EvaluateEnsemble
