+++ src/ert/gui/summarypanel.py
