+++ src/ert/gui/tools/plot/data_type_keys_widget.py
@@ -1,15 +1,71 @@
 from typing import Dict, List, Optional
 
-from qtpy.QtCore import Signal
-from qtpy.QtGui import QIcon
-from qtpy.QtWidgets import QHBoxLayout, QListView, QToolButton, QVBoxLayout, QWidget
+from qtpy.QtCore import QSize, Signal
+from qtpy.QtGui import QColor, QIcon, QPainter, QPaintEvent
+from qtpy.QtWidgets import (
+    QHBoxLayout,
+    QLabel,
+    QListView,
+    QToolButton,
+    QVBoxLayout,
+    QWidget,
+)
 
-from ert.gui.ertwidgets import Legend, SearchBox
+from ert.gui.ertwidgets import SearchBox
 from ert.gui.tools.plot.plot_api import PlotApiKeyDefinition
 
 from .data_type_keys_list_model import DataTypeKeysListModel
 from .data_type_proxy_model import DataTypeProxyModel
 from .filter_popup import FilterPopup
+
+
+class _LegendMarker(QWidget):
+    """A widget that shows a colored box"""
+
+    def __init__(self, color: QColor):
+        QWidget.__init__(self)
+
+        self.setMaximumSize(QSize(12, 12))
+        self.setMinimumSize(QSize(12, 12))
+
+        self.color = color
+
+    def paintEvent(self, a0: Optional[QPaintEvent]) -> None:
+        painter = QPainter(self)
+
+        rect = self.contentsRect()
+        rect.setWidth(rect.width() - 1)
+        rect.setHeight(rect.height() - 1)
+        painter.drawRect(rect)
+
+        rect.setX(rect.x() + 1)
+        rect.setY(rect.y() + 1)
+        painter.fillRect(rect, self.color)
+
+
+class _Legend(QWidget):
+    """Combines a _LegendMarker with a label"""
+
+    def __init__(self, legend: Optional[str], color: QColor):
+        QWidget.__init__(self)
+
+        self.setMinimumWidth(140)
+        self.setMaximumHeight(25)
+
+        self.legend = legend
+
+        layout = QHBoxLayout()
+        layout.setContentsMargins(0, 0, 0, 0)
+
+        self.legend_marker = _LegendMarker(color)
+        self.legend_marker.setToolTip(legend)
+
+        layout.addWidget(self.legend_marker)
+        self.legend_label = QLabel(legend)
+        layout.addWidget(self.legend_label)
+        layout.addStretch()
+
+        self.setLayout(layout)
 
 
 class DataTypeKeysWidget(QWidget):
@@ -49,7 +105,7 @@ class DataTypeKeysWidget(QWidget):
         layout.addStretch()
 
         layout.addWidget(
-            Legend("Observations available", DataTypeKeysListModel.HAS_OBSERVATIONS)
+            _Legend("Observations available", DataTypeKeysListModel.HAS_OBSERVATIONS)
         )
 
         self.setLayout(layout)
