+++ src/ert/resources/workflows/jobs/internal-gui/scripts/csv_export.py
@@ -161,10 +161,7 @@ class CSVExportJob(ErtPlugin):
         return export_info
 
     def getArguments(self, parent, ert_config, storage):
-        from ert.gui.ertwidgets.customdialog import CustomDialog
-        from ert.gui.ertwidgets.listeditbox import ListEditBox
-        from ert.gui.ertwidgets.models.path_model import PathModel
-        from ert.gui.ertwidgets.pathchooser import PathChooser
+        from ert.gui.ertwidgets import CustomDialog, ListEditBox, PathChooser, PathModel
 
         description = "The CSV export requires some information before it starts:"
         dialog = CustomDialog("CSV Export", description, parent)
