+++ tests/unit_tests/gui/tools/test_workflow_tool.py
@@ -8,7 +8,7 @@ import pytest
 from qtpy.QtCore import Qt, QTimer
 
 from ert.config import ErtConfig
-from ert.gui.ertwidgets.closabledialog import ClosableDialog
+from ert.gui.ertwidgets import ClosableDialog
 from ert.gui.main import _setup_main_window
 from ert.gui.main_window import ErtMainWindow
 from ert.gui.tools.event_viewer import GUILogHandler
